moved to indicator package
