+++ Makefile
@@ -7,7 +7,7 @@ dev:
 	PYTHONPATH=. uv run python jesse/rl/main.py
 
 test:
-	uv run python -m pytest tests/test_rl_strategy.py::TestRLStrategy::test_rl_strategy
+    uv run python -m pytest tests/test_rl_strategy.py::TestRLStrategy::test_rl_strategy
 
 run:
     uv run jesse run
