+++ jesse/indicators/rl/profit_reward_double_dqn_model
