+++ jesse/indicators/rl/sr_reward_double_ddqn_model
